requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
colorama>=0.4.6
